requirements.txt
# Core ML
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0
datasets>=2.19.0
accelerate>=0.30.0
peft>=0.10.0

# Training & Experiment Tracking
mlflow>=2.12.0
optuna>=3.6.0
ray[tune]>=2.10.0
wandb>=0.17.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
pyarrow>=15.0.0
scikit-learn>=1.4.0
scipy>=1.13.0

# Feature Store
feast>=0.38.0

# Vector Database
faiss-cpu>=1.8.0
weaviate-client>=4.5.0

# Serving
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
httpx>=0.27.0

# RAG / LLM
langchain>=0.2.0
langchain-community>=0.2.0
openai>=1.30.0
tiktoken>=0.7.0
sentence-transformers>=2.7.0

# Streaming / Messaging
kafka-python>=2.0.2
aiokafka>=0.10.0

# Monitoring & Observability
evidently>=0.4.30
prometheus-client>=0.20.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0

# Storage & Infra
boto3>=1.34.0
google-cloud-storage>=2.16.0
redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
typer>=0.12.0
rich>=13.7.0
loguru>=0.7.0
tenacity>=8.3.0
tqdm>=4.66.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
httpx>=0.27.0
factory-boy>=3.3.0
